feat(stat): rewrite select queries in dsl api

+++ stat-service/src/main/kotlin/com/fleetmate/stat/modules/car/dto/CarListItemDto.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class CarListItemDto(
     val car: CarSimpleDto,
     val fuelLevel: Double,
-    var violations: Int = 0
+    var violations: Long = 0L
 )
